Typos fixed, and blank lines added
Note that I'm not a native English speaker, so I might just as well have it wrong!

The added blank lines hopefully result in small comment boxes appearing for logged-in users:
I noticed that some list items have them popping up next to them, and some don't. I suspect it's because of a preceding blank line.